* Remove unnecessary type conversions * Remove unnecessart version information * gofmt

+++ rpcserver.go
@@ -575,7 +575,7 @@ func handleCreateRawTransaction(s *rpcServer, cmd interface{}, closeChan <-chan 
 			return nil, internalRPCError(err, context)
 		}
 
-		txOut := wire.NewTxOut(int64(satoshi), pkScript)
+		txOut := wire.NewTxOut(satoshi, pkScript)
 		mtx.AddTxOut(txOut)
 	}
 
@@ -1167,7 +1167,7 @@ func handleGetBlock(s *rpcServer, cmd interface{}, closeChan <-chan struct{}) (i
 	var retargetEstimate *float64
 
 	blocksBetweenRetargets := int32(params.TargetTimespan / params.TargetTimePerBlock)
-	blocksUntilRetarget := blocksBetweenRetargets - (int32(blockHeight) % blocksBetweenRetargets)
+	blocksUntilRetarget := blocksBetweenRetargets - (blockHeight % blocksBetweenRetargets)
 	if blocksUntilRetarget == blocksBetweenRetargets {
 		blocksUntilRetarget = 0
 	}
